Move component OpenDataLink from 'components' to 'components/_shared'
